[refactor] 패키지 리펙토링
