--- app/src/main/java/org/android/go/sopt/data/local/ApiFactory.kt
@@ -1,33 +1,0 @@
-package org.android.go.sopt.data.local
-
-import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
-import kotlinx.serialization.json.Json
-import okhttp3.MediaType.Companion.toMediaType
-import retrofit2.Retrofit
-
-object ApiFactory {
-    val retrofit: Retrofit by lazy {
-        Retrofit.Builder()
-            .baseUrl("http://52.78.152.187:8080/")
-            .addConverterFactory(Json.asConverterFactory("application/json".toMediaType()))
-            .build()
-    }
-    inline fun <reified T> create(): T = retrofit.create<T>(T::class.java)
-}
-
-object ReqresApiFactory {
-    val Reqretrofit: Retrofit by lazy {
-        Retrofit.Builder()
-            .baseUrl("https://reqres.in/")
-            .addConverterFactory(Json.asConverterFactory("application/json".toMediaType()))
-            .build()
-    }
-    inline fun <reified T> create(): T = Reqretrofit.create<T>(T::class.java)
-}
-
-object ServicePool {
-    val signUpService = ApiFactory.create<SignUpService>()
-    val logInService = ApiFactory.create<LogInService>()
-    val followersService = ReqresApiFactory.create<FollowersService>()
-
-}
