--- app/src/main/java/org/android/go/sopt/data/local/FollowersDto.kt
@@ -1,32 +1,0 @@
-package org.android.go.sopt.data.local
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ResponseFollowersDto(
-    @SerialName("page")
-    val page: Int,
-    @SerialName("per_page")
-    val per_page: Int,
-    @SerialName("total")
-    val total: Int,
-    @SerialName("total_pages")
-    val total_pages: Int,
-    @SerialName("data")
-    val data: List<RequestFollowersData>,
-){
-    @Serializable
-    data class RequestFollowersData(
-        @SerialName("id")
-        val id: Int,
-        @SerialName("email")
-        val email: String,
-        @SerialName("first_name")
-        val first_name: String,
-        @SerialName("last_name")
-        val last_name: String,
-        @SerialName("avatar")
-        val avatar: String,
-    )
-}
