--- app/src/main/java/org/android/go/sopt/data/local/SignUpService.kt
@@ -1,11 +1,0 @@
-package org.android.go.sopt.data.local
-
-import retrofit2.Call
-import retrofit2.http.Body
-import retrofit2.http.POST
-
-interface SignUpService {
-    @POST("sign-up")
-    fun signUp(@Body request: RequestSignUpDto
-    ): Call<ResponesSignUpDto>
-}
