+++ app/src/main/java/org/android/go/sopt/data/response/ResponseLogInDto.kt
@@ -1,21 +1,10 @@
-package org.android.go.sopt.data.local
+package org.android.go.sopt.data.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RequestSignUpDto(
-    @SerialName("id")
-    val id: String,
-    @SerialName("password")
-    val password: String,
-    @SerialName("name")
-    val name: String,
-    @SerialName("skill")
-    val skill: String
-)
-@Serializable
-data class ResponesSignUpDto(
+data class ResponseLogInDto(
     @SerialName("status")
     val status: Int,
     @SerialName("message")
@@ -25,9 +14,13 @@ data class ResponesSignUpDto(
 ){
     @Serializable
     data class UserInfo(
+        @SerialName("id")
+        val id: String,
         @SerialName("name")
         val name: String,
         @SerialName("skill")
-        val skill: String,
+        val skill: String
     )
 }
+
+
