+++ app/src/main/java/org/android/go/sopt/presentation/gallery/adapter/GalleryAdapter.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.adapter
+package org.android.go.sopt.presentation.gallery.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
