+++ app/src/main/java/org/android/go/sopt/presentation/home/view/HomeFragment.kt
@@ -1,0 +1,16 @@
+package org.android.go.sopt.presentation.home.view
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import org.android.go.sopt.databinding.FragmentHomeBinding
+import org.android.go.sopt.data.api.ServicePool
+import org.android.go.sopt.presentation.home.adapter.FriendAdapter
+import retrofit2.Call
+import retrofit2.Response
+/*실제 사용자에게 보여지는 화면*/
+class HomeFragment : Fragment() {
+
+}
