+++ app/src/main/res/layout/activity_login.xml
@@ -59,7 +59,7 @@
         app:layout_constraintTop_toBottomOf="@id/et_text_inputid" />
 
     <EditText
-        android:id="@+id/et_text_inputps"
+        android:id="@+id/et_text_inputpw"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginHorizontal="30dp"
@@ -81,7 +81,7 @@
         android:text="@string/login"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/et_text_inputps" />
+        app:layout_constraintTop_toBottomOf="@+id/et_text_inputpw" />
 
     <Button
         android:id="@+id/bt_signIn"
