+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.login.LoginActivity">
+    tools:context=".presentation.login.view.LoginActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fcv_main"
