+++ app/src/main/res/layout/item_friend.xml
